g4-common: Cleanup, remove deprecated TARGET_RIL_VARIANT
hardware/ril is called in the build repo. its forced to build with it, once we remove the TARGET_RIL_VARIANT parameter.
This fixes our RIL crashes, since we no longer build OSS RIL, but it is built by the ROM (using hardware/ril).

This is thanks to changes that Google have made in Pie..

Change-Id: I4d94532b5c55e60f5e41704e3c2545bda7f33e2e